create generic trainer and modify mtad-gat training
